Split participant cost tests into their own file
test_reporting.py had grown past CodeScene's function-count threshold
for a single module. TestCostsByParticipant follows the existing
test_reporting_filters.py precedent of splitting reporting tests by
concern into a dedicated file.
